+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 2.2.4
 ++++++
+* `vmss create`: reserve large enough frontend port range for overprovision
 * `az sig`: fix update commands, support --no-wait on managing image versions
 * `vm list-ip-addresses`: now shows availability zone of public IP addresses.